Tighten infeasibility tolerance for MIP
When running the experimental pipeline on the demo datasets, our assertion
for modelling strict inequalities was violated. In particular, the
difference between feature values in at least one dataset was lower than
the infeasibility tolerance that Python-MIP uses to accept inqualities as
satsified. To make sure the assertion is always satisified, we set the
infeasibility tolerance to 0 now (potentially increasing the runtime).