Use auxiliary in-box variables in SMT formulation
Technically unnecessary, as the corresponding Boolean expressions can be
inlined, like they were before this commit. In particular, each of the new
Boolean variables has an equivalence (<->) correspondence to a Boolean
expression in the objective. However, these auxiliary variables allow to
extract the box's bounds in the same way as for the MIP formulation: Use
the minimum/maximum feature values of instances in the box. In the MIP
case, this (1) prevented numerical issues (instances classified as in-box
by optimizer could be slightly out-of-box if comparing feature values to
the bounds) and (2) made the box "tight" around the contained instances
(i.e., bounds assume exact feature values instead of spreading into the
"margin").
For SMT, we did not encounter such numeric issues yet, and the boxes were
tight by default , but since the auxiliary variables do not seem to hurt
runtime performance, we include them anyway (for consistency reasons and
to counter potential numeric issues in the future).