Adapt evaluation to competitors and timeout-support of pipeline
Changes introduced in 65237215c and 9a9aa1dca3.
The competitors are not a problem per se (we already compared MIP and SMT
as SD methods), but as the number of SD methods has increased considerably
(2 -> 8), results may need a different format if printed to the console.
The timeout either becomes another dimension in the evaluation or is fixed
(to its maximum) in other evaluations.

New evaluations:
- impact of timeout on evaluation metrics
- impact of timeout on overfitting
- correlation of runtime with number of unique feature values