Re-implement prelim's PRIM with feature-cardinality support
Could have also created a subclass and just overridden _peel_one(), but to
get better control over the implementation and make changes here and there
(in particular, termination condition and interpolation of quantiles; in
the future, may also add a pasting routine), we re-implemented it.
As a result, the BI algorithm (which we still wrap from "prelim") gets its
own class now, since it's the only algorithm from "prelim" we use (i.e., a
more general wrapper is unnecessary).