Re-implement beam search with cardinality support
pysubgroup's BeamSearch class has a parameter for the search depth, but
that is not the number of used features because
1) a lower and an upper bound for one feature already are two conditions
2) conditions may be redundant (e.g., a lower bound for one feature may
   be replaced with a tighter lower bound later, which increases the length
   of the subgroup description / search depth but does not constitute the
   use of a new feature)

Therefore, we re-implement beam search to allow cardinality constraints as
we have for all other our algorithms. The new routine also is considerably
faster. Without cardinality constraints, the achieved WRAcc is nearly or
exactly the same. Since BeamSearch was the last algorithm from pysubgroup
we used (others were buggy or terribly slow), we remove this package and
its dependencies from our environment definition altogether.

We also refactor the cardinality handling in PRIM, but that should not
affect the results (we only move one computation into an if-statement
where it is needed).