Add alternative subgroup descriptions for SMT
Besides optimizing WRacc, can now also optimize similarity to the optimal
subgroup description s.t. using different features. Both optimization
problems share many variables and constraints, so we created a joint
optimization method. fit() dispatches to this method and returns the same
result as before. We adapt evaluate() for SMT to be able to return
alternative descriptions. We also adapt the return type of evaluate() in
the superclass and the experimental pipeline for compatibility reasons.
The search parameters for alternative subgroup descriptions are not part
of the pipeline yet, i.e., the pipeline should work and yield the same
results as before.