Fix feature cardinality count for exact optimizers
In the heuristics for subgroup-discovery, we iteratively try out different
candidate values for the bounds. We directly use the best bound candidates
as actual bounds, only replacing feature minima/maxima with -/+inf.
In contrast, MIP and SMT use a more indirect procedure to alleviate numeric
issues: While these optimization problems have decision variables for the
bounds, we do not use them as final bounds. Instead, we used the decision
variables telling which instances are in the box. Based on these variables,
we chose the minimum and maximum feature values of instances in the box.
This procedure is problematic when we want to determine the number of
features used. For example, a feature may not need to be actively
restricted by the optimizer, but (due to restrictions on other features)
its minimum and maximum value in the box may not correspond to its global
minimum and maximum. Thus, based on the chosen bounds, more features may
appear restricted than actually were in optimization, and the bounds for
these features may be tighter than necessary (though they still yield the
same prediction on the training data and the feature-cardinality constraint
during optimization is observed). This behavior becomes a larger problem
when searching for alternative subgroup descriptions, because there we
deduce feature usage based on the finally set bounds, not usage in
optimization.
Solution: Introduce new variables whether lower or upper bound are active
(we only had less specific variables whether feature used, i.e., either
bound active) and use -/+inf for all inactive bounds.