Normalize WRAcc in evaluation
The maximum value of WRAcc depends not only on how the instances are
distributed (which may make predictions easier or harder), but also the
target distribution. In particular, the more imbalanced the classes are,
the smaller the effective range of WRAcc is, even if a perfect box (all
positives, no negatives) could be found. E.g., even for perfectly balanced
classes, WRAcc only has the range [-0.25, 0.25], while for 90:10, it has
the range [-0.09, 0.09]. As the PLMB datasets have different degrees of
class imbalance, we normalize WRAcc in the evaluation to [-1, 1] by
dividing by its maximum value to give the different datasets equal
influence on the mean WRAcc.
In the future, we may directly use the normalized score in the evaluation
function called in the pipeline.