Adapt normalization of alternatives evaluation
Originally (9449c87b98), we used plain max normalization. However, that
has weaknesses if the quality is close to zero (large ratios or even div by
zero) or negative (which may particularly happen on the test set). Also,
the maximum may not only be at the original subgroup but also any
alternative, so the point of reference is not fixed. Thus, we adapt the
evaluation as follows:

1) Shift range of evaluation metrics from [-1, 1] to [0, 1] (as we did in
  AFS evaluation), so values close to zero are very unlikely (random
  guessing yields 0.5).
2) Divide by quality of original subgroup instead of max-quality.