Introduce superclass for search for alternative
When we first added the search for alternatives in 4d92d003ce, it was
limited to one subgroup-discovery method (SMT). However, we want to
implement this functionality for at least one other SD method. Further, the
search loop over alternatives and evaluation of alternatives are
independent from the SD method anyway. Thus, we extract this general code
to a new abstract superclass (which inherits from the top-level abstract
superclass) and only leave the optimization routine in the subclass.
This commit should not affect the functionality of the SMT subclass at all.