Add alternative subgroup descriptions to beam search
After implementing this functionality for SMT first(4d92d003ce) and adding
an abstract superclass later (1b638449e6), we now enhance beam search with
the option to search for alternative subgroup descriptions. This search
problem requires

- a new objective function (Hamming distance instead of WRAcc), which we
  implemented before (aae6f2240f) and now activate if alternatives desired
- further constraints which features may be used in candidate boxes
  (force deselection from existing subgroups); we extract these constraints
  into a new method, also moving the cardinality checks there

As a refactoring, we also extract the cardinality constraints for PRIM and
BestInterval in a similar manner.

Additionally, we changed the notion of alternatives from "select >= tau_abs
new features" to "deselect >= tau_abs old features" since the latter
formulation is antimonotonic (while the former is not) and therefore better
suited for greedy forward search. We adapt the definition of alternatives
in SMT optimization accordingly.