Add normalized WRAcc to pipeline
WRACC was already normalized in evaluation (post-hoc, i.e., after running
the experiments) since 2e8a00aff5. In this commit, we add normalized WRAcc
to the pipeline, storing it besides the "ordinary" WRAcc. We only need to
add this new metric to evaluate() in the abstract superclass for subgroup
discovery since WRAcc only differs by a constant from WRAcc, so the
optimization objectives of the subgroup-discovery methods can remain as-is.