Adapt handling of empty boxes in PRIM peeling
Make code consistent to paper, though behavior should not change.
Before, we used a boolean flag to indicate whether peeling produced an
empty box. However, we did not update the box's bounds, so one absolutely
had to check this flag in the stop condition of PRIM's main loop since the
number of instances in the box and predictions were wrong.
With this refactoring, we remove the boolean flag and instead produce an
empty (invalid) box from peeling, which causes PRIM to stop automatically
(unless the support threshold is negative ...). Optimizing WRAcc, this will
never be the solution of PRIM overall, as it has the same quality as the
unrestriced box, with which PRIM starts.