Fix initialization of PRIM bounds
Instead of using -/+ inf like in beam search, we started from the feature
minima/maxima. In theory, this shouldn't make a difference, as the initial
subgroup still contains all data objects. In practice, this means the
subgroup's bounds may be stored in int arrays instead of float arrays if
the dataset only contains integers. That is not a problem per se, but
numpy's int arrays also cast each newly assigned value to int, even if the
value is float. This could be problematic when determining quantiles within
PRIM (which may fall between feature values), but I guess they were just
converted to float as well. What actually caused the code to crash on
certain datasets was the assignment of infinite bounds for empty boxes
(since commit 9da2992388). While we also regularly assign +/- inf in the
post-processing of many subgroup-discovery methods, theses assignments work
with pd.Series instead of np.arrays. The former automatically change the
array type from int to float if necessary to accommodate a float value.