Add evaluation of actual vs desired feature cardinality
As the experimental pipeline stores the feature sets (since 7b5b377188),
we can easily count how many features are actually selected and compare
this to "k" (which only is an upper bound, not the exact number). To this
end, we need to convert the feature lists stored as strings in the CSV back
to proper lists. We also noted that the same conversion routine does not
directly work for the lower/upper bounds, since these lists also contain
+/-inf; however, we don't require the latter columns in our current
evaluation, so we didn't implement a fix.