Dynamically initialize qualities for Beam and BI
Before this commit, the quality of all subgroups in the beam was
initialized to 0. This is the correct value for WRAcc as the objective and
assuming unrestricted subgroups (which we use indeed). However, for an
evaluation metric other than WRAcc, 0 may not be correct. In particular,
BeamSearch also supports searching alternative subgroup descriptions.
There, unrestricted subgroups will have some overlap with the original
subgroup (unless the former is empty) and therefore start with a non-zero
quality. Thus, we actively compute the initial quality for BeamSearch and
BestInterval now. The results for searching original subgroups should not
be affected at all, while there may be minor changes in the search for
alternatives (however, it is unlikely the best subgroup after search will
be an initial subgroup anyway).