Fix NaN founds in solver-based search for empty subgroups
In solver-based search, we do not use the bound values from the
corresponding numeric variables but the minimum/maximum of all data objects
in the box instead, which (1) makes the box as tight as possible (not
necessarily required, but consistent to our other subgroup-discovery
methods) and (2) avoids numeric issues when extracting the bounds from the
solver (in the past, there were corner cases with disagreement between the
box-membership variables from the solver and computing box membership with
the extracted bounds).
However, computing the minimum/maximum requires data objects to be in the
subgroup. This is not always the case, particularly with solver timeouts
and when searching for alternative subgroup descriptions, where the empty
subgroup may have similar membership as another subgroup containing few
data objects. For empty subgroups, bounds were saved as NaN. Funnily, the
feature selection was deduced correctly from this (NaN is not infinity)
and the predictions with these bounds were probably correct as well
(comparing NaN to an actual number yields false). Nevertheless, we fix
this issue by only computing the min/max if possible and returning invalid
infinite bounds (LB = +inf, UB = -inf) for empty subgroups.