User proper docstrings for subgroup-discovery methods
Already documented the evaluation metrics in 74d3f71ee1. Use numpy-style
docstrings. Increases the number of lines in the file by roughly 60%.
All functions and classes in the Python package "csd" are properly
documented now. Minor changes to some type hints and parameter order in
some methods, which should not affect the functionality of the code.