Add no-timeout cardinality plots for diss evaluation
In the (current) arXiv version of the paper, there are two plots for the
impact of feature cardinality "k" (i.e., train and test set), both using
all datasets, i.e., including timeouts. For the dissertation, we create
exactly the same plot types also for just the datasets without timeout.
Further, we set the "style_order" in this plot type to make the mapping of
line styles to SD methods consistent (as the hue order).