Re-organize experimental code
The repo's top-level became a bit cluttered after adding the runtime
experiments. Thus, we move main and runtime experiments into separate
folders. The only code they share is the "csd" package (separate anyway)
and "data_handling.py" (which we copy). Further, we describe the files
and reproduction steps of the competitor-runtime experiments in the README.
Finally, we delete the (unused) demo-dataset preparation script.